requirements.txt
Copypdf2image==1.16.3
pytesseract==0.3.10
Pillow==10.0.0
PyPDF2==3.0.1
python-logging-config==1.0.3